Adds an Evolution class and the ability to record acceptance ratios.
* The Evolution class records the history of the DNA.
* We can now use a finite value of P, which is the intrinsic disorder
  persistence length. If I understand correctly, P is basically the length
  scale of "straightness" at zero force. This creates a correction in B,
  the new microscopic value is called Bm as in Bryan's paper.
* The "extension" calculation has been changed so all the three coordinates
  of the bead are reported, not just the z coordinate.

We could possibly use macros somehow (c.f. macropy) to reduce the excessive
code duplication between the subclasses of NakedDNA. I think that would be
better instead of having a 'god' function and you pass it everything and
some flags to tweak behaviour.

I also realized that I should use reStructuredText and not markdown syntax
for formatting the docstrings...