Adds functionality for easily having multiple runs of arbitrary simulations.
Data is now saved as an xarray Dataset instead of a plain dictionary. One may
think of Datasets as fancy dictionaries where the values are numpy arrays and
the indices can be labeled using coordinates. Datasets allow easier
aggregration of data, similar to Pandas dataframes. The xarray docs describe
why Datasets (and DataArrays) are easier to use over Pandas dataframes.

Using Datasets makes it easy to have multiple runs of the same simulation and
accumulate the results easily, merging common parameters as appropriate.
The main work is done by concat_datasets in the utils module.
See the (_)simulate_dna functions in ipython_utils as a usage example.