Bug fix for stretching energy + refactoring code for angular description.
The stretching energy computation assumed that the tangent vector was scaled
but actually it wasn't. I've renamed the related functions (and we should
maintain this in the future) so that the function clearly indicates if the
result contains only unit vectors.

We have a common class with several useful functions for describing the
angles in sim_utils. This is shared between the molecular dynamics and the
Monte Carlo simulations, avoiding unnecessary code duplication.

* The Jacobian test is failing now because we added nonzero terms as per
  Bryan's note (Eq. C6) for the ψ row.
* Adds lots of docstrings and comments for easier debugging later on.
* Fixed code style in several places, according to PEP 8.
* Variable renaming for better readability.