Changes boundary conditions in MC to match with MD.
It makes more sense logically to have fixed orientation conditions at the ends
of the rod chain. The bottom of the first rod is euler[0] - this is fixed.
The bottom of the last rod is euler[-1] - this is _not_ fixed.
Instead, its top is fixed (attribute 'end'), which interacts with euler[-1] to
give an additional energy term, so the last rod can move.

* Replaced 1D force in stretch_energy_density with 3D.
* Put env attribute (has temperature) inside AngularDescription.
* Edited docs for clarity in some places.
* Added a script so we can save .ipynb files as well in the future.
* Fixed randomization bug in metropolis function.
* Use boolean reject values as that is more intuitive.
* Changed some function names according to PEP 8.